+++ src/aui-modal/js/aui-modal.js
@@ -59,13 +59,11 @@ A.Modal = A.Base.create('modal', A.Widget, [
         eventHandles = [
             A.after(instance._afterFillHeight, instance, 'fillHeight'),
             instance.after('render', instance._afterRender),
-            instance.after('resize:end', A.bind(instance._syncResizeDimensions, instance)),
             instance.after('draggableChange', instance._afterDraggableChange),
-            instance.after('resizableChange', instance._afterResizableChange),
             instance.after('visibleChange', instance._afterVisibleChange)
         ];
 
-        instance._applyPlugin(instance._onUserInitInteraction);
+        instance._applyPlugin(instance._plugDrag);
 
         instance._eventHandles = eventHandles;
     },
@@ -144,24 +142,6 @@ A.Modal = A.Base.create('modal', A.Widget, [
     },
 
     /**
-     * Fire after resize changes.
-     *
-     * @method _afterResizableChange
-     * @param event
-     * @protected
-     */
-    _afterResizableChange: function(event) {
-        var instance = this;
-
-        if (event.newVal) {
-            instance._applyPlugin(instance._plugResize);
-        }
-        else {
-            instance.unplug(A.Plugin.Resize);
-        }
-    },
-
-    /**
      * Fire after visibility changes.
      *
      * @method _afterVisibleChange
@@ -226,40 +206,18 @@ A.Modal = A.Base.create('modal', A.Widget, [
     },
 
     /**
-     * Fire before resizing to the correct dimensions.
-     *
-     * @method _beforeResizeCorrectDimensions
-     * @param event
-     * @protected
-     */
-    _beforeResizeCorrectDimensions: function() {
-        var instance = this;
-
-        if (instance.resize.proxy) {
-            return new A.Do.Prevent();
-        }
-    },
-
-    /**
-     * Plug draggable/resizable if enable.
+     * Plug draggable and detach user interaction handle.
      *
      * @method _onUserInitInteraction
      * @protected
      */
     _onUserInitInteraction: function() {
-        var instance = this,
-            draggable = instance.get('draggable'),
-            resizable = instance.get('resizable');
+        var instance = this;
 
+        instance._plugDrag();
+
+        instance._userInteractionHandle.detach();
         instance._userInteractionHandle = null;
-
-        if (draggable) {
-            instance._plugDrag();
-        }
-
-        if (resizable) {
-            instance._plugResize();
-        }
     },
 
     /**
@@ -272,47 +230,9 @@ A.Modal = A.Base.create('modal', A.Widget, [
         var instance = this,
             draggable = instance.get('draggable');
 
-        instance.plug(A.Plugin.Drag, instance._addBubbleTargets(draggable));
-    },
-
-    /**
-     * Plug the resize Plugin
-     *
-     * @method _plugResize
-     * @protected
-     */
-    _plugResize: function() {
-        var instance = this,
-            resizable = instance.get('resizable');
-
-        instance.plug(A.Plugin.Resize, instance._addBubbleTargets(resizable));
-
-        A.before(instance._beforeResizeCorrectDimensions, instance.resize, '_correctDimensions', instance);
-    },
-
-    /**
-     * Sync width/height dimensions on resize.
-     *
-     * @method _syncResizeDimensions
-     * @param event
-     * @protected
-     */
-    _syncResizeDimensions: function(event) {
-        var instance = this,
-            boundingBox = instance.get('boundingBox'),
-            resize = event.info;
-
-        instance.set(
-            'width',
-            resize.offsetWidth -
-            parseInt(boundingBox.getComputedStyle('borderRightWidth'), 10) -
-            parseInt(boundingBox.getComputedStyle('borderLeftWidth'), 10));
-
-        instance.set(
-            'height',
-            resize.offsetHeight -
-            parseInt(boundingBox.getComputedStyle('borderTopWidth'), 10) -
-            parseInt(boundingBox.getComputedStyle('borderBottomWidth'), 10));
+        if (draggable) {
+            instance.plug(A.Plugin.Drag, instance._addBubbleTargets(draggable));
+        }
     }
 }, {
 
@@ -377,19 +297,6 @@ A.Modal = A.Base.create('modal', A.Widget, [
                         fn: A.Plugin.DDConstrained
                     }
                 ]
-            }
-        },
-
-        /**
-         * Determine if Modal should be resizable or not.
-         *
-         * @attribute resizable
-         * @type Object
-         * @writeOnce
-         */
-        resizable: {
-            value: {
-                handles: 'br'
             }
         },