+++ src/aui-modal/HISTORY.md
@@ -8,5 +8,6 @@ No registries yet.
 
 ## [2.5.0](https://github.com/liferay/alloy-ui/releases/tag/2.5.0)
 
+* [AUI-1306](https://issues.liferay.com/browse/AUI-1306) Make aui-modal responsive
 * [AUI-1163](https://issues.liferay.com/browse/AUI-1163) Remove unnecessary constants
 * [AUI-1107](https://issues.liferay.com/browse/AUI-1107) Listeners after visibleChange are not executed with modal.destroyOnHide enabled